+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-04-14 20:31:08
-// Commit: chore(ui): update dependencies and cleanup unused imports in ui
+// Updated at: 2026-04-16 15:05:55
+// Commit: perf(ui): apply RepaintBoundary on animated ui widgets
 
 class uiComponentState {
-  final String commitRef = '379';
-  final String lastMessage = 'chore(ui): update dependencies and cleanup unused imports in ui';
-  final DateTime timestamp = DateTime.parse('2026-04-14T20:31:08+05:00');
+  final String commitRef = '425';
+  final String lastMessage = 'perf(ui): apply RepaintBoundary on animated ui widgets';
+  final DateTime timestamp = DateTime.parse('2026-04-16T15:05:55+05:00');
 }